+++ CounterWebApp/src/main/resources/spring/batch/jobs/job-report.xml
@@ -55,10 +55,9 @@
 					<bean
 						class="org.springframework.batch.item.file.transform.DelimitedLineTokenizer">
 						<property name="names"
-							value="date,impressions,clicks,earning" />
+							value="creditCardNumber" />
 					</bean>
 				</property>
-
 				<property name="fieldSetMapper">
 
 					<!-- return back to reader, rather than a mapped object. -->
@@ -84,8 +83,8 @@
 		<property name="sql">
 			<value>
             <![CDATA[        
-            	insert into RAW_REPORT(REPORT_DATE,IMPRESSIONS,CLICKS,EARNING, JOB_ID, JOB_START_TIME, PROCESSING_FILE, ACQUIRER, CHUNK_COMMIT_SIZE  ) values
-            	 (:date, :impressions, :clicks, :earning, :jobId, :loadedDate, :processingFile, :acquirer, :commitSize)
+            	insert into ACQUIRER_SETTLEMENTS(CREDIT_CARD_NUMBER, META_JOB_ID, META_JOB_START_TIME, META_PROCESSING_FILE, META_ACQUIRER, META_CHUNK_COMMIT_SIZE  ) values
+            	 (:creditCardNumber, :jobId, :loadedDate, :processingFile, :acquirer, :commitSize)
             ]]>
 			</value>
 		</property>
